# 우클릭 문제 해결

+++ Code/Protocol/Packet.h
@@ -284,7 +284,7 @@ struct C2S_InputPacket : public PacketUdpHeader {
 	float    MoveX = 0.0f; // -1~1
 	float    MoveY = 0.0f; // -1~1
 	float	 MoveZ = 0.0f; // -1~1
-	uint8    Buttons = 0;  // 비트플래그 (점프/발사/대시 등)
+	uint32    Buttons = 0;  // 비트플래그 (점프/발사/대시 등)
 	float    Yaw = 0.0f;
 	float    Pitch = 0.0f;
